Move RunWithConfig to internal app package
